Исправлена сборка и улучшена работа rasl-constants на Linux (#77)
Во-первых, скрипт make.sh теперь имеет флаг исполнимости.

Во-вторых, в compiler/makeself{,-s}.sh указано правильное имя файла.

В-третьих, путь к ../srlib/refalrts.h в compiler/makeself{,-s}.bat указан
с прямыми слешами — в шапке комментария он отображается одинаково
на Windows и Linux.

В-четвёртых, исправлена опечатка в шапке генеримого комментария.

В-пятых, src/make.* не выдаёт сообщения об ошибке об отсутствии файла
common/*.cpp.

+++ src/make.sh
@@ -57,7 +57,7 @@ make_subdir() {
 
     mkdir -p ../../build/$DIR
     mv *.cpp ../../build/$DIR
-    mv ../common/*.cpp ../../build/$DIR
+    [ -e  ../common/*.cpp ] && mv ../common/*.cpp ../../build/$DIR
     cp $PATH_TO_SREFC/srlib/*.cpp ../../build/$DIR
   fi
 )
